requirements.txt
aioredis==1.2.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
certifi==2019.3.9
cffi==1.12.2
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
daphne==2.2.5
deap==1.2.2
Django==2.1.7
django-rest-knox==4.0.0
djangorestframework==3.9.2
h2==3.1.0
hiredis==1.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
msgpack-python==0.5.6
numpy==1.16.2
pandas==0.24.2
priority==1.3.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2018.9
redis==2.10.6
requests==2.22.0
rest-pandas==1.1.0
scikit-learn==0.20.3
scipy==1.2.1
service-identity==18.1.0
six==1.12.0
stopit==1.1.2
TPOT==0.9.6
tqdm==4.31.1
Twisted==18.9.0
txaio==18.8.1
update-checker==0.16
urllib3==1.24.2
whitenoise==4.1.2
zope.interface==4.6.0
psycopg2
tseries_featurizer==0.5
django-rest-swagger
numba==0.43.1
gunicorn==19.9.0
